% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARTall.R
\name{STARTall}
\alias{STARTall}
\title{Evaluates the imported patients' data for all the START criteria at once.}
\usage{
STARTall(path, export_data_path)
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from.}

\item{export_data_path}{(Character) (optional) (default: working directory) the path for excel file output.}
}
\description{
Evaluates the imported patients' data for all the START criteria at once.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
