% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARTselected.R
\name{STARTselected}
\alias{STARTselected}
\title{Evaluates the imported patients' data for the selected START criteria at once.}
\usage{
STARTselected(selected, path, export_data_path)
}
\arguments{
\item{selected}{(Character vector) the selected criteria to export given in the form of c("A1", "E2")}

\item{path}{(Character) the path that the excel file can be read from.}

\item{export_data_path}{(Character) (optional) (default: working directory) the path for excel file output.}
}
\description{
Evaluates the imported patients' data for the selected START criteria at once.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
