% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARTSTOPPbycategory.R
\name{STARTSTOPPbycategory}
\alias{STARTSTOPPbycategory}
\title{Evaluates the imported patients' data for all the START and STOPP criteria at once and returns the results by category (overuse, misuse, underuse) and not by each criterion (e.g. STOPP B1, STOPP B2, etc.).}
\usage{
STARTSTOPPbycategory(
  path = NULL,
  STARTselected = "all",
  STARTexclude = NULL,
  STOPPselected = "all",
  STOPPexclude = NULL,
  excel_out = TRUE,
  single_excel = TRUE,
  export_data_path = NULL,
  suppressNA = TRUE
)
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from.}

\item{STARTselected}{(Character) (optional) (default: c('all')) a vector of START criteria that you want to include. Example: c("B3", "C1").}

\item{STARTexclude}{(Character) (optional) (default: NULL) a vector of START criteria that you want to exclude. Example: c("B3", "C1").}

\item{STOPPselected}{(Character) (optional) (default: c('all')) a vector of STOPP criteria that you want to include Example: c("C6", "D2").}

\item{STOPPexclude}{(Character) (optional) (default: NULL) a vector of STOPP criteria that you want to exclude. Example: c("C6", "D2").}

\item{excel_out}{(Boolean) (optional) (default: TRUE) output excel file with the evaluated data.}

\item{single_excel}{(Boolean) (optional) (default: TRUE) if true outputs only 1 excel file with multiple columns instead of multiple files (one for each criterion)}

\item{export_data_path}{(Character) (optional) (default: NULL (a popup message to choose dir will be displayed)) the path for excel file output.}

\item{suppressNA}{(Boolean) (optional) (default: TRUE) set this to FALSE if you want to know for which patients have NAs and for which variable. By default all NAs will be ignored so that the algorithm can distinguish between patients who meet the criterion and those who do not.}
}
\description{
Evaluates the imported patients' data for all the START and STOPP criteria at once and returns the results by category (overuse, misuse, underuse) and not by each criterion (e.g. STOPP B1, STOPP B2, etc.).
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
