% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOPP_F4.R
\name{STOPP_F4}
\alias{STOPP_F4}
\title{Evaluates the imported patients' data for the STOPP F4 criterion.}
\usage{
STOPP_F4(path, excel_out = TRUE, export_data_path = getwd())
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from.}

\item{excel_out}{(Boolean) (optional) (default: TRUE) output excel file with the evaluated data.}

\item{export_data_path}{(Character) (optional) (default: working directory) the path for excel file output.}
}
\value{
list of lists of evaluated patient ids categorized in 1) the ids that fulfill the criterion, 2) the ids that do not fulfill the criterion and 3) the ids that has missing data
}
\description{
Evaluates the imported patients' data for the STOPP F4 criterion.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
