% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOPPall.R
\name{STOPPall}
\alias{STOPPall}
\title{Evaluates the imported patients' data for the all the STOPP criteria at once.}
\usage{
STOPPall(path, export_data_path)
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from.}

\item{export_data_path}{(Character) (optional) (default: working directory) the path for excel file output.}
}
\description{
Evaluates the imported patients' data for the all the STOPP criteria at once.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
