% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOPPselected.R
\name{STOPPselected}
\alias{STOPPselected}
\title{Evaluates the imported patients' data for the selected STOPP criteria at once.}
\usage{
STOPPselected(selected, path, export_data_path)
}
\arguments{
\item{selected}{(Character vector) the selected criteria to export given in the form of c("B1", "D5", "H3")}

\item{path}{(Character) the path that the excel file can be read from.}

\item{export_data_path}{(Character) (optional) (default: working directory) the path for excel file output.}
}
\description{
Evaluates the imported patients' data for the selected STOPP criteria at once.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
