% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOPPselected.R
\name{STOPPselected}
\alias{STOPPselected}
\title{Evaluates the imported patients' data for the selected STOPP criteria at once.}
\usage{
STOPPselected(
  selected = NULL,
  exclude = NULL,
  path = NULL,
  excel_out = TRUE,
  single_excel = TRUE,
  export_data_path = NULL,
  suppressNA = TRUE,
  returnData = FALSE
)
}
\arguments{
\item{selected}{(Character vector) the selected criteria to export given in the form of c("A1", "E2")}

\item{exclude}{(Character) (optional) (default: NULL) a vector of criteria that you want to exclude. Example: c("C6", "D2").}

\item{path}{(Character) the path that the excel file can be read from.}

\item{excel_out}{(Boolean) (optional) (default: TRUE) output excel file with the evaluated data.}

\item{single_excel}{(Boolean) (optional) (default: TRUE) if true outputs only 1 excel file with multiple columns instead of multiple files (one for each criterion)}

\item{export_data_path}{(Character) (optional) (default: NULL (a popup message to choose dir will be displayed)) the path for excel file output.}

\item{suppressNA}{(Boolean) (optional) (default: TRUE) set this to FALSE if you want to know for which patients have NAs and for which variable. By default all NAs will be ignored so that the algorithm can distinguish between patients who meet the criterion and those who do not.}

\item{returnData}{(Boolean) (optional) (default: FALSE) set this to TRUE if you want to have data returned in a data frame format.}
}
\value{
NULL or data.frame. Returns the evaluated data for each criterion if you set the returnData argument to TRUE. By default it does not return anything.
}
\description{
Evaluates the imported patients' data for the selected STOPP criteria at once.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
